mcp-servers/memory-mcp/requirements.txt
fastapi
uvicorn
pydantic
psycopg2-binary
prometheus-fastapi-instrumentator>=6.1.0
